refactor: split helpers.ts into focused modules
helpers.ts had grown to 1620 lines mixing unrelated concerns. Extract
three cohesive groups into their own files and re-export them from
helpers.ts (barrel pattern), so no call sites change:

- colors.ts      - hex/HSV/rgb conversions
- cssUtils.ts    - px/spacing/background/unit normalization
- scriptSandbox.ts - user block-script execution (restricted + unrestricted)

helpers.ts: 1620 -> 1009 lines.
